refactor: 알림 권한 동기화 로직을 앱 활성화 시마다 수행하도록 리팩토링
