+++ app/(tabs)/_layout.tsx
@@ -1,11 +1,99 @@
 import { Tabs } from 'expo-router';
-import React from 'react';
-import { SafeAreaView, Image, Text, StatusBar, Platform } from 'react-native';
-import { Dimensions } from 'react-native';
+import React, { useEffect, useRef, useState } from 'react';
+import { Image, Text, AppState, AppStateStatus, Dimensions, Alert } from 'react-native';
+import messaging from '@react-native-firebase/messaging';
+import * as Notifications from 'expo-notifications';
+import {
+  postFcmDeviceToken,
+  getNotificationsSettingStatus,
+  initNotificationsSettingStatus,
+  putOSPushAgreement,
+} from '@/api/notifications/notifications';
+import NotificationPermissionModal from '@/components/NotificationPermissionModal';
+import queryClient from '@/api/queryClient';
+
 const { height: screenHeight } = Dimensions.get('window');
 const TAB_BAR_HEIGHT = screenHeight * 0.117; // 화면 높이의 15%
 
 export default function TabLayout() {
+  const appState = useRef<AppStateStatus>(AppState.currentState);
+  const [isNotificationPermissionModalOpen, setIsNotificationPermissionModalOpen] = useState(false);
+
+  /* 1. FCM 토큰 등록 */
+  const updateFcmToken = async () => {
+    await messaging().registerDeviceForRemoteMessages();
+    const token = await messaging().getToken();
+    await postFcmDeviceToken(token);
+  };
+
+  /* 2. OS 권한 요청 및 서버 동기화 */
+  const initOSPermissionStatus = async () => {
+    try {
+      const authStatus = await messaging().requestPermission();
+      const enabled =
+        authStatus === messaging.AuthorizationStatus.AUTHORIZED ||
+        authStatus === messaging.AuthorizationStatus.PROVISIONAL;
+      await putOSPushAgreement(enabled);
+
+      setIsNotificationPermissionModalOpen(false);
+
+      if (!enabled) {
+        Alert.alert('Notifications are turned off', 'You can enable notifications for Kori in Settings.');
+        return;
+      }
+
+      await initNotificationsSettingStatus(true);
+    } catch (error) {
+      console.error('[ERROR] Notification permission or sync failed:', error);
+    }
+  };
+
+  /* 3. 앱 초기화 시마다 OS 알림 권한을 검사 및 서버와 동기화 */
+  useEffect(() => {
+    const handleAppStateChange = async (nextState: AppStateStatus) => {
+      const isActivated =
+        (appState.current.match(/inactive|background/) && nextState === 'active') || // 백그라운드에서 포그라운드 진입 시
+        (appState.current === 'active' && nextState === 'active'); // 앱 최초 실행 시
+
+      if (isActivated) {
+        try {
+          await updateFcmToken();
+
+          const { status } = await Notifications.getPermissionsAsync();
+
+          if (status === 'undetermined') {
+            // 아직 권한 설정 전인 경우 설정 요청 모달 표시
+            setIsNotificationPermissionModalOpen(true);
+            return;
+          }
+
+          if (status === 'denied') {
+            // OS 권한이 거부된 경우 서버 알림 전체 비활성화
+            await initNotificationsSettingStatus(false);
+          } else if (status === 'granted') {
+            // OS 권한이 허용된 경우 서버 상태 점검
+            const serverStatus = await getNotificationsSettingStatus();
+            if (serverStatus === 'NEEDS_SETUP') {
+              // 서버 설정이 필요한 경우 OS 상태 동기화 여부와 알림 상세 설정을 모두 true로 초기화
+              await putOSPushAgreement(true);
+              await initNotificationsSettingStatus(true);
+            }
+          }
+          await queryClient.invalidateQueries({ queryKey: ['notificationSettings'] });
+        } catch (error) {
+          console.error('[ERROR] 알림 권한 갱신 중 오류 발생:', error);
+        }
+      }
+
+      appState.current = nextState;
+    };
+
+    const subscription = AppState.addEventListener('change', handleAppStateChange);
+    handleAppStateChange('active');
+
+    return () => subscription.remove();
+  }, []);
+
   return (
     <Tabs
       screenOptions={{
@@ -103,6 +191,18 @@ export default function TabLayout() {
               resizeMode="contain"
             />
           ),
+        }}
+      />
+      {/* 알림 권한 모달 */}
+      <NotificationPermissionModal
+        visible={isNotificationPermissionModalOpen}
+        onClose={() => setIsNotificationPermissionModalOpen(false)}
+        onYesPress={async () => {
+          await initOSPermissionStatus();
+        }}
+        onLaterPress={async () => {
+          await putOSPushAgreement(false);
+          await initNotificationsSettingStatus(false);
         }}
       />
     </Tabs>
