+++ app/(tabs)/index.tsx
@@ -1,13 +1,5 @@
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import {
-  ActivityIndicator,
-  Alert,
-  AppState,
-  AppStateStatus,
-  DeviceEventEmitter,
-  FlatList,
-  RefreshControl,
-} from 'react-native';
+import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import { ActivityIndicator, Alert, DeviceEventEmitter, FlatList, RefreshControl } from 'react-native';
 import styled from 'styled-components/native';
 
 import FriendCard from '@/components/FriendCard';
@@ -19,17 +11,7 @@ import { useSentFollowRequestsSet } from '@/hooks/queries/useFollowList';
 import useMyProfile from '@/hooks/queries/useMyProfile';
 import useRecommendedFriends from '@/hooks/queries/useRecommendedFriends';
 import { router } from 'expo-router';
-import {
-  getNotificationsSettingStatus,
-  initNotificationsSettingStatus,
-  postFcmDeviceToken,
-  putOSPushAgreement,
-} from '@/api/notifications/notifications';
-import messaging from '@react-native-firebase/messaging';
-import NotificationPermissionModal from '@/components/NotificationPermissionModal';
-import { useNotificationStore } from '@/src/store/useNotificationStore';
 import { Text } from '@react-navigation/elements';
-import { getOSNotificationPermissionStatus } from '@/lib/fcm/getOSPermissionStatus';
 
 const toBirthNumber = (v: unknown): number | undefined => {
   if (typeof v === 'number') return Number.isFinite(v) ? v : undefined;
@@ -38,7 +20,6 @@ const toBirthNumber = (v: unknown): number | undefined => {
 };
 
 export default function HomeScreen() {
-  const appState = useRef<AppStateStatus>(AppState.currentState);
   const { data: friends, isLoading, isFetching, refetch } = useRecommendedFriends(20);
   const followMutation = useFollowUser();
   const cancelReqMutation = useCancelFollowRequest();
@@ -47,8 +28,6 @@ export default function HomeScreen() {
 
   const [requested, setRequested] = useState<Set<number>>(new Set());
   const [inFlight, setInFlight] = useState<Set<number>>(new Set());
-  const [isNotificationPermissionModalOpen, setIsNotificationPermissionModalOpen] = useState<boolean>(false);
-  const { isNotificationNeedsSetup, setIsNotificationNeedsSetup } = useNotificationStore();
 
   const lock = (id: number) => setInFlight((s) => new Set(s).add(id));
   const unlock = (id: number) =>
@@ -128,65 +107,6 @@ export default function HomeScreen() {
       next.delete(id);
       return next;
     });
-  }, []);
-
-  /* 1. FCM 토큰 발급 및 업데이트 */
-  const updateFcmToken = async () => {
-    await messaging().registerDeviceForRemoteMessages();
-    const token = await messaging().getToken();
-    await postFcmDeviceToken(token);
-  };
-
-  /* 2. 알림 설정 상태 확인 */
-  const checkNotificationStatus = async () => {
-    const serverStatus = await getNotificationsSettingStatus();
-    const osStatus = await getOSNotificationPermissionStatus();
-
-    const needsSetupFromServer = serverStatus === 'NEEDS_SETUP';
-    const needsSetupFromOS = osStatus === 'undetermined';
-
-    const needsSetup = needsSetupFromServer || needsSetupFromOS;
-
-    if (needsSetup) {
-      setIsNotificationPermissionModalOpen(true);
-    } else {
-      updateOSPermissionStatus();
-    }
-  };
-
-  /* 3. OS 권한 요청 및 서버 동기화 */
-  const updateOSPermissionStatus = async (needsToInitStatus: boolean = false) => {
-    try {
-      const authStatus = await messaging().requestPermission();
-      const enabled =
-        authStatus === messaging.AuthorizationStatus.AUTHORIZED ||
-        authStatus === messaging.AuthorizationStatus.PROVISIONAL;
-      await putOSPushAgreement(enabled);
-
-      setIsNotificationPermissionModalOpen(false);
-
-      if (!enabled) {
-        Alert.alert('Notifications are turned off', 'You can enable notifications for Kori in Setting.');
-        setIsNotificationNeedsSetup(false);
-        return;
-      }
-
-      if (needsToInitStatus) {
-        await initNotificationsSettingStatus(true);
-      }
-      setIsNotificationNeedsSetup(false);
-    } catch (error) {
-      console.error('[ERROR] 알림 권한 요청 및 서버 동기화 실패:', error);
-    }
-  };
-
-  useEffect(() => {
-    if (isNotificationNeedsSetup) {
-      (async () => {
-        await updateFcmToken();
-        await checkNotificationStatus();
-      })();
-    }
   }, []);
 
   return (
@@ -284,18 +204,6 @@ export default function HomeScreen() {
           }
         />
       )}
-      {/* 알림 허용 모달 */}
-      <NotificationPermissionModal
-        visible={isNotificationPermissionModalOpen}
-        onClose={() => setIsNotificationPermissionModalOpen(false)}
-        onYesPress={async () => {
-          await updateOSPermissionStatus(true);
-        }}
-        onLaterPress={async () => {
-          await putOSPushAgreement(false);
-          await initNotificationsSettingStatus(false);
-        }}
-      />
     </Safe>
   );
 }
