--- lib/fcm/getOSPermissionStatus.ts
@@ -1,6 +1,0 @@
-import * as Notifications from 'expo-notifications';
-
-export const getOSNotificationPermissionStatus = async () => {
-  const { status } = await Notifications.getPermissionsAsync();
-  return status;
-};
